Updated tileset 37.
Major changes to tileset 37. It looks like the original developers
changed this tileset after creating the Dodongo Caverns, but didn't
re-do the dungeon with the new tileset. I fixed that by updating this
tileset. It requires a bit of work/map editing, but the bulk of the
dungeon is intact and this should update it visually.