Bug fixes and hacking fixes.
Banks were so broken! SO BROKEN! If you put a 0 as the amount, it would
deposit the item but let you keep it. Same thing for taking an item from
the bank. What a simple oversight.
I ran through the code and looked for a few more simple < 0 bugs.
ParseCommand for admins now shouldn't crash the server.. instead you get
an error message.